+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 
+## [1.2.0] - 2017-10-19
+
+## Added
+- FaceID support (fingerprintLock changed to biometricLock)
+
+## [1.1.1] - 2017-10-19
+
+## Fixed
+- Migration from custom Locker implementation to Locker v1 (for iOS version < 9)
+
 ## [1.1.0] - 2017-10-05
 
 ## Added